fix: BadResource ID in custom install, unittest in one file
